Backend Fixed Image

+++ app/Http/Controllers/UserProfileController.php
@@ -122,7 +122,11 @@ class UserProfileController extends Controller
         $userInfo->facebook = $request->facebook;
         $userInfo->twitter = $request->twitter;
         $userInfo->instagram = $request->instagram;
-        unlink(realpath($userInfo->image));
+        //return dump( realpath( $userInfo->image));
+        if ($request->image!=null){
+            if (realpath( $userInfo->image))
+            unlink( realpath( $userInfo->image) );
+        }
         $userInfo->image = $request->hasFile('image') ? $this->uploadFile($request->file('image')) : "defaultImage.png";
         if ($userInfo->save())
             // return redirect()->route('list_categories')->with(['success' => 'تم تحديث البيانات بنجاح']);
